requirements_dev.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
aleph-alpha-client==2.17.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
blinker==1.8.2
boto3==1.35.39
botocore==1.35.39
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.3
chromadb==0.4.14
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
deeplake==3.9.26
dill==0.3.9
distro==1.9.0
exceptiongroup==1.2.2
faiss-cpu==1.9.0
fastapi==0.115.2
ffmpeg==1.4
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
gpt-index==0.4.15
greenlet==3.1.1
grpcio==1.66.2
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.2
humanfriendly==10.0
humbug==0.3.2
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.10
langchain-text-splitters==0.3.0
langsmith==0.1.134
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==3.0.1
marshmallow==3.22.0
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
mutagen==1.47.0
mypy-extensions==1.0.0
narwhals==1.9.3
nltk==3.9.1
openai==0.28.0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pathos==0.3.3
pillow==10.4.0
posthog==3.7.0
pox==0.3.5
ppft==1.7.6.9
propcache==0.2.0
protobuf==5.28.2
pulsar-client
pyarrow==17.0.0
pycares==4.4.0
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
PyPDF2==3.0.1
PyPika==0.48.9
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==12.1.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
s3transfer
safetensors
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==1.4.54
starlette==0.39.2
streamlit==1.39.0
sympy==1.13.3
tenacity==8.1.0
tiktoken
tokenizers
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
watchdog==5.0.3
watchfiles==0.24.0
websockets==13.1
yarl==1.15.0
yt-dlp
